[2120] Clear bursary info when changing publish course
The bursary info can change if your course changes. We previously added
this to the course details form, but your course can also change in the
confirm publish course form.
