+++ app/forms/concerns/course_form_helpers.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module CourseFormHelpers
+  def clear_bursary_information
+    trainee.progress.funding = false
+
+    trainee.assign_attributes({
+      applying_for_bursary: nil,
+    })
+  end
+
+  def course_subjects_changed?
+    trainee.course_subject_one_changed? || trainee.course_subject_two_changed? || trainee.course_subject_three_changed?
+  end
+end
